sync main

+++ builder.sh
@@ -3,30 +3,39 @@ set -e
 set -o pipefail
 
 main() {
-	mkdir --parents /out/cache/custom/{src,pkg} /out/cache/pkg /home/custompkgs
-	mv /out/cache/pkg/* /var/cache/pacman/pkg/. || true 
+	mkdir --parents /out/cache/custom/{src,pkg} /out/cache/pkg
+	mv /out/cache/pkg/* /var/cache/pacman/pkg/. || true
 	
 	pacman-key --init
 	pacman --sync --refresh --noconfirm archlinux-keyring
-	pacman --sync --refresh --sysupgrade --noconfirm --needed git pacman-contrib
+	pacman --sync --refresh --sysupgrade --noconfirm --needed git pacman-contrib opendoas
 	git config --global --add safe.directory /packages
 
+	# makepkg hack to run as root
+ 	sed 's,exit $E_ROOT,#exit $E_ROOT,' --in-place /usr/bin/makepkg
+  	sed "s,'verifysource' 'version','verifysource' 'version' 'asroot'," --in-place /usr/bin/makepkg
+
+  	echo "permit nopass :archie as root cmd /usr/bin/mkosi" > /etc/doas.conf
+   	chown -c root:root /etc/doas.conf
+    	chmod -c 0400 /etc/doas.conf
+
 	useradd --create-home archie
-	chown --recursive archie /out /home/custompkgs /packages
+	chown --recursive archie /out /packages
 	echo "archie ALL=(ALL) NOPASSWD: /usr/bin/pacman" > "/etc/sudoers.d/allow_archie_to_pacman"
 	echo "root ALL=(ALL) CWD=* ALL" > /etc/sudoers.d/permissive_root_Chdir_Spec
 
-	rm -rf /home/custompkgs/*
-	runuser -u archie -- repo-add /home/custompkgs/custom.db.tar.gz
-	find /out/cache/custom/pkg -type f -name '*.pkg.tar.zst' -exec runuser -u archie -- repo-add /home/custompkgs/custom.db.tar.gz {} \;
+	if test ! -f /out/cache/custom/pkg/custom.db.tar.gz; then
+		runuser -u archie -- repo-add --new /out/cache/custom/pkg/custom.db.tar.gz
+  	fi
+	find /out/cache/custom/pkg -type f -name '*.pkg.tar.zst' -exec runuser -u archie -- repo-add --new /out/cache/custom/pkg/custom.db.tar.gz {} \;
 
 	if ! grep 'custom.conf' /etc/pacman.conf; then
 		echo "Include = /etc/pacman.d/custom.conf" >> /etc/pacman.conf
 	fi
-	cat <<-'EOF' > "/etc/pacman.d/custom.conf"
+	cat <<- 'EOF' > /etc/pacman.d/custom.conf
 		[custom]
 		SigLevel = Optional TrustAll
-		Server = file:///home/custompkgs
+		Server = file:///out/cache/custom/pkg
 	EOF
 	echo 'PKGDEST=/out/cache/custom/pkg' > /etc/makepkg.conf.d/pkgdest.conf
 	echo 'SRCPKGDEST=/out' > /etc/makepkg.conf.d/srcpkgdest.conf
@@ -42,27 +51,42 @@ main() {
  	runuser -u archie -- makepkg-url "https://aur.archlinux.org/cgit/aur.git/plain/{PKGBUILD,aurutils.changelog,aurutils.install}?h=aurutils" --syncdeps --install --clean --noconfirm --rmdeps
 
  	echo "Cache is $(ls /out/cache/custom/pkg)"
+  	#echo 1 > /proc/sys/kernel/unprivileged_userns_clone
 
-	tehbuild() {
+	tehbuildloop() {
 		cd "${1}"
 		if test -f PKGBUILD; then
 			if ! grep '^# do not build' PKGBUILD; then
+   				if test ! -f /tmp/fail; then
 				echo "Considering $(basename "$(pwd)")"
-				for f in $(runuser -u archie -- makepkg --packagelist); do
-    					echo "Looking for ${f}"
-    					if test -f "${f}"; then
-	 					echo "We already had ${f}"
-       						ls -al /out/cache/custom/pkg
-       					else
-						echo "Building $(basename "$(pwd)")"
-						runuser -u archie -- paru --upgrade --noconfirm
-						ln -s ./cache/custom/pkg/$(basename "${f}") /out/.
-						runuser -u archie -- makepkg --allsource  # --sign
-      						#mv *.pkg.tar.zst.sig /out/.
-						echo "Done building $(basename "$(pwd)")"
-      						break
-	 				fi
-				done
+					for f in $(runuser -u archie -- makepkg --packagelist); do
+	    					echo "Looking for ${f}"
+	    					if test -f "${f}"; then
+		 					echo "We already had ${f}"
+	       						ls -al /out/cache/custom/pkg
+	       					else
+							echo "Building $(basename "$(pwd)")"
+       							if test "$(basename "$(pwd)")" = "alarm_image_pi5"; then
+	      							#runuser -u archie -- paru -Syu --needed --noconfirm archlinuxarm-keyring btrfs-progs cpio f2fs-tools dosfstools erofs-utils mtools squashfs-tools pixz
+	      							#makepkg --asroot -Cfi
+	      							runuser -u archie -- paru --upgrade --noconfirm
+	      						else
+								runuser -u archie -- paru --upgrade --noconfirm
+							fi
+		    					if test -f "${f}"; then
+								echo "Done building $(basename "$(pwd)")"
+		       						ln -s ./cache/custom/pkg/$(basename "${f}") /out/.
+								runuser -u archie -- makepkg --allsource  # --sign
+		      						#mv *.pkg.tar.zst.sig /out/.
+	       						else
+		     						touch /tmp/fail
+		    					fi
+	      						break
+		 				fi
+					done
+     				else
+					echo "Skipping $(pwd) because of precious failure"
+  				fi
 			else
 				echo "Skipping $(pwd) because # do not build in PKGBUILD"
 			fi
@@ -70,9 +94,14 @@ main() {
 			echo "Skipping $(pwd) because no PKGBUILD"
 		fi
 	}
-	
-	export -f tehbuild
-	find /packages/ -maxdepth 1 -type d -exec bash -c 'tehbuild "${0}"' "{}" \;
+	export -f tehbuildloop
+ 
+	find /packages/ -maxdepth 1 -type d -exec bash -c 'tehbuildloop "${0}"' "{}" \;
+ 	if test -f /tmp/fail; then
+  		echo "ERROR: Couldn't find ${f} after building it."
+    		exit -44
+      	fi
+
 	git clean -ffxd || true
 	paccache --remove --keep 1
  	paccache --remove --keep 1 --min-mtime "1 day ago" --cachedir /out/cache/custom/pkg
